requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cutlet==0.1.19
docker==6.1.2
frozenlist==1.4.0
fugashi==1.3.0
idna==3.4
jaconv==0.3.4
mojimoji==0.0.12
multidict==6.0.4
musicbrainzngs==0.7.1
mutagen==1.46.0
mypy-extensions==1.0.0
openai==0.28.0
packaging==23.1
Pillow==9.5.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
requests==2.28.2
soupsieve==2.4
tabulate==0.9.0
tqdm==4.66.1
typing_extensions==4.9.0
unidic-lite==1.0.8
urllib3==1.26.15
wcwidth==0.2.9
websocket-client==1.5.1
yarl==1.9.2
